Start farmer protocol, PeerConnections object, full node server

+++ src/types/challenge.py
@@ -1,10 +1,11 @@
-from ..util.streamable import streamable
-from .sized_bytes import bytes32
+from src.util.streamable import streamable
+from src.types.sized_bytes import bytes32
+from src.util.ints import uint32, uint64
 
 
 @streamable
 class Challenge:
     proof_of_time_output_hash: bytes32
     proof_of_space_hash: bytes32
-    height: int
-    total_weight: int
+    height: uint32
+    total_weight: uint64
